various changes to make the repo self contained
