+++ .gitignore
@@ -36,8 +36,8 @@ pip-delete-this-directory.txt
 # Unit test / coverage reports
 htmlcov/
 .tox/
-.coverage
 .coverage.*
+.coverage
 .cache
 nosetests.xml
 coverage.xml
