+++ capiq_client/tests/unit/test_capiq_client.py
@@ -1,6 +1,6 @@
 from mock import mock
 import unittest
-from capiq_client import CapIQClient
+from capiq_client.capiq_client import CapIQClient
 
 orig_import = __import__
