+++ capiq_client/tests/unit/test_capiq_client_gdshv.py
@@ -1,8 +1,8 @@
 from mock import mock
 import unittest
-from capiq_client import CapIQClient
+from capiq_client.capiq_client import CapIQClient
 
-def mocked_gdsg_data_requests_post(*args, **kwargs):
+def mocked_gdshv_data_requests_post(*args, **kwargs):
     class MockResponse:
         def __init__(self, json_data, status_code):
             self.json_data = json_data
@@ -10,11 +10,6 @@ def mocked_gdsg_data_requests_post(*args, **kwargs):
 
         def json(self):
             return self.json_data
-    """
-    if args[0] == 'http://someurl.com/test.json':
-        return MockResponse({"key1": "value1"}, 200)
-    elif args[0] == 'http://someotherurl.com/anothertest.json':
-    """
     if args[0] is not None:
         return MockResponse({"GDSSDKResponse": [      {
             "Headers": ["IQ_CLOSEPRICE"],
@@ -22,7 +17,7 @@ def mocked_gdsg_data_requests_post(*args, **kwargs):
             "NumCols": 1,
             "Seniority": "",
             "Mnemonic": "IQ_CLOSEPRICE",
-            "Function": "GDSG",
+            "Function": "GDSHV",
             "ErrMsg": None,
             "Properties": {},
             "StartDate": "",
@@ -34,7 +29,7 @@ def mocked_gdsg_data_requests_post(*args, **kwargs):
             "Limit": ""
         }]}, 200)
 
-def mocked_gdsg_no_data_requests_post(*args, **kwargs):
+def mocked_gdshv_no_data_requests_post(*args, **kwargs):
     class MockResponse:
         def __init__(self, json_data, status_code):
             self.json_data = json_data
@@ -54,7 +49,7 @@ def mocked_gdsg_no_data_requests_post(*args, **kwargs):
                             "NumCols": 1,
                             "Seniority": "",
                             "Mnemonic": "IQ_CLOSEPRICE",
-                            "Function": "GDSG",
+                            "Function": "GDSHV",
                             "ErrMsg": "SOME ERROR",
                             "Properties": {},
                             "StartDate": "",
@@ -68,16 +63,28 @@ def mocked_gdsg_no_data_requests_post(*args, **kwargs):
                     ]
             }, 200)
 
-class TestCapiqClientGdsg(unittest.TestCase):
+class TestCapiqClientGdshv(unittest.TestCase):
 
-    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdsg_data_requests_post)
-    def test_gdsg_data(self, mocked_post):
+    @mock.patch('capiq_client.capiq_client.requests.post', side_effect=mocked_gdshv_data_requests_post)
+    def test_gdshv_data(self, mocked_post):
         ciq_client = CapIQClient()
-        return_value = ciq_client.gdsg(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], properties=[{}])
+        return_value = ciq_client.gdshv(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], properties=[{}])
         self.assertEqual(return_value, {'TRIP:': {'close_price': '46.80'}})
 
-    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdsg_no_data_requests_post)
-    def test_gdsg_no_data(self, mocked_post):
+    @mock.patch('capiq_client.capiq_client.requests.post', side_effect=mocked_gdshv_no_data_requests_post)
+    def test_gdst_no_data(self, mocked_post):
         ciq_client = CapIQClient()
-        return_value = ciq_client.gdsg(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
+        return_value = ciq_client.gdshv(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
         self.assertEqual(return_value, {'TRIP:': {'close_price': None}})
+
+    @mock.patch('capiq_client.capiq_client.requests.post', side_effect=mocked_gdshv_data_requests_post)
+    def test_gdst_data_no_properties(self, mocked_post):
+        ciq_client = CapIQClient()
+        return_value = ciq_client.gdshv(
+            ["TRIP"],
+            ["IQ_CLOSEPRICE"],
+            ["close_price"],
+            start_date="12/19/1980",
+            end_date="12/19/2000"
+        )
+        self.assertEqual(return_value, {'TRIP:': {'close_price': '46.80'}})
