+++ capiq_client/tests/unit/test_capiq_client_gdsp.py
@@ -1,8 +1,8 @@
 from mock import mock
 import unittest
-from capiq_client import CapIQClient
+from capiq_client.capiq_client import CapIQClient
 
-def mocked_gdshe_data_requests_post(*args, **kwargs):
+def mocked_gdsp_data_requests_post(*args, **kwargs):
     class MockResponse:
         def __init__(self, json_data, status_code):
             self.json_data = json_data
@@ -10,14 +10,15 @@ def mocked_gdshe_data_requests_post(*args, **kwargs):
 
         def json(self):
             return self.json_data
+
     if args[0] is not None:
-        return MockResponse({"GDSSDKResponse": [      {
+        return MockResponse({"GDSSDKResponse": [{
             "Headers": ["IQ_CLOSEPRICE"],
             "Rows": [{"Row": ["46.80"]}],
             "NumCols": 1,
             "Seniority": "",
             "Mnemonic": "IQ_CLOSEPRICE",
-            "Function": "GDSHE",
+            "Function": "GDSP",
             "ErrMsg": None,
             "Properties": {},
             "StartDate": "",
@@ -29,7 +30,7 @@ def mocked_gdshe_data_requests_post(*args, **kwargs):
             "Limit": ""
         }]}, 200)
 
-def mocked_gdshe_no_data_requests_post(*args, **kwargs):
+def mocked_gdsp_no_data_requests_post(*args, **kwargs):
     class MockResponse:
         def __init__(self, json_data, status_code):
             self.json_data = json_data
@@ -37,7 +38,11 @@ def mocked_gdshe_no_data_requests_post(*args, **kwargs):
 
         def json(self):
             return self.json_data
-
+    """
+    if args[0] == 'http://someurl.com/test.json':
+        return MockResponse({"key1": "value1"}, 200)
+    elif args[0] == 'http://someotherurl.com/anothertest.json':
+    """
     if args[0] is not None:
         return MockResponse(
             {
@@ -49,7 +54,7 @@ def mocked_gdshe_no_data_requests_post(*args, **kwargs):
                             "NumCols": 1,
                             "Seniority": "",
                             "Mnemonic": "IQ_CLOSEPRICE",
-                            "Function": "GDSHE",
+                            "Function": "GDSP",
                             "ErrMsg": "SOME ERROR",
                             "Properties": {},
                             "StartDate": "",
@@ -63,28 +68,16 @@ def mocked_gdshe_no_data_requests_post(*args, **kwargs):
                     ]
             }, 200)
 
-class TestCapiqClientGdshe(unittest.TestCase):
+class TestCapiqClientGdsp(unittest.TestCase):
 
-    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdshe_data_requests_post)
-    def test_gdshe_data(self, mocked_post):
+    @mock.patch('capiq_client.capiq_client.requests.post', side_effect=mocked_gdsp_data_requests_post)
+    def test_gdsp_data(self, mocked_post):
         ciq_client = CapIQClient()
-        return_value = ciq_client.gdshe(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], properties=[{}])
+        return_value = ciq_client.gdsp(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
         self.assertEqual(return_value, {'TRIP:': {'close_price': '46.80'}})
 
-    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdshe_no_data_requests_post)
-    def test_gdshe_no_data(self, mocked_post):
+    @mock.patch('capiq_client.capiq_client.requests.post', side_effect=mocked_gdsp_no_data_requests_post)
+    def test_gdsp_no_data(self, mocked_post):
         ciq_client = CapIQClient()
-        return_value = ciq_client.gdshe(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
+        return_value = ciq_client.gdsp(["TRIP"], ["IQ_CLOSEPRICE"], ["close_price"], [{}])
         self.assertEqual(return_value, {'TRIP:': {'close_price': None}})
-
-    @mock.patch('capiq_client.requests.post', side_effect=mocked_gdshe_data_requests_post)
-    def test_gdst_data_no_properties(self, mocked_post):
-        ciq_client = CapIQClient()
-        return_value = ciq_client.gdshe(
-            ["TRIP"],
-            ["IQ_CLOSEPRICE"],
-            ["close_price"],
-            start_date="12/19/1980",
-            end_date="12/19/2000"
-        )
-        self.assertEqual(return_value, {'TRIP:': {'close_price': '46.80'}})
